requirements.txt
numpy==1.22.3
pandas==1.4.3
pandas_stubs==1.2.0.56
Pillow==9.2.0
requests==2.27.1
streamlit==1.11.1
tensorflow==2.9.1
gTTS==2.2.4